Remove +1, add if else statements
Completely unnecessary, thought it was a good idea at the time for some
reason. -_- Forgot to commit, but stored the answer of Math.floor(Math.
random()*3) into x, and added if else statement to make each number
equal "Rock", "Paper", or "Scissors" and return those strings in place
of the numbers.