Make bunch of changes, forgot to commit dammit
Declared playerInput to store the result of the prompt to choose which,
then declared playerChoice to store the answer to the prompt but all
lowercased, then declare computerChoice as a variable which will store
the result of the function getComputerChoice().

Wrote if statement saying if the result was equal to "rock", "paper",
or "scissors, then run playRound(playerChoice, computerChoice), then
store the result inside another variable called gameResult.

Then return the value of gameResult.